Create Github Action to automatically build and publish to NPM when a new release is created
This sets up a Github Action that triggers every time a new release is created in GH. It runs tests, builds, updates the
package.json version to the one defined by the tag in the Release, then publishes to NPM.

This makes the process of creating a release super simple and only one step: Publish a release with the desired version
in Github.

Test Plan: Create a new Release in the Github UI with a tag that is one version higher than the current version in package.json.
Ensure this workflow is triggered and that the new version is released to NPM and the package.json version is updated as well on master.