add cache-clearing step in load cli command
This is a relatively naive implementation of a cache clearing retry for
the load command. It works as follows:
1. check all plugins to be loaded for an existing cache db and save the
state to a Map
2. For each plugin if the load fails for any reason and a cache exists,
  the cache is deleted and the load is retried. Otherwise the load fails
  as before.
3. If the single retry fails, the load will fail.

Some alternative implementations were considered but decided against for
the following reasons:
- the CLI context has no knowledge of the db file name composition
semantics, so naively checking the cache dir was implemented in favor of
looking for a specific file
- It's difficult to catalog all reasons the cache may be
invalidated across plugins, so this implementation does not parse the
error type in order to determine whether a retry is the proper next step

test plan:
1. run `path/to/sourcred.js load` to set up an initital cache
2. perform an action that would cause a sqlite db error or a
graphql error using the master cli load command
3. run `path/to/sourcred.js load` again to verify the error does not
occur. Instead, the cache db should be deleted and recreated